refactor: separation of exceptions
