+++ src/main/java/ulaval/glo2003/Main.java
@@ -10,8 +10,8 @@ import org.glassfish.jersey.server.ResourceConfig;
 import ulaval.glo2003.api.HealthResource;
 import ulaval.glo2003.api.product.ProductResource;
 import ulaval.glo2003.api.seller.SellerResource;
-import ulaval.glo2003.api.exceptionHandling.NotFoundExceptionMapper;
-import ulaval.glo2003.api.exceptionHandling.ParamExceptionMapper;
+import ulaval.glo2003.api.exceptionMappers.NotFoundExceptionMapper;
+import ulaval.glo2003.api.exceptionMappers.ParamExceptionMapper;
 import ulaval.glo2003.domain.Seller;
 
 public class Main {
