+++ src/main/java/ulaval/glo2003/api/exceptionMappers/NotFoundExceptionMapper.java
@@ -1,8 +1,9 @@
-package ulaval.glo2003.api.exceptionHandling;
+package ulaval.glo2003.api.exceptionMappers;
 
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.ExceptionMapper;
+import ulaval.glo2003.domain.exceptions.ErrorDescription;
 
 public class NotFoundExceptionMapper implements ExceptionMapper<NotFoundException> {
