+++ src/main/java/ulaval/glo2003/api/exceptionMappers/ParamExceptionMapper.java
@@ -1,7 +1,8 @@
-package ulaval.glo2003.api.exceptionHandling;
+package ulaval.glo2003.api.exceptionMappers;
 
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.ExceptionMapper;
+import ulaval.glo2003.domain.exceptions.ParamValidationException;
 
 public class ParamExceptionMapper implements ExceptionMapper<ParamValidationException> {
