+++ src/main/java/ulaval/glo2003/domain/Product.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import ulaval.glo2003.api.product.ProductCategory;
-import ulaval.glo2003.api.exceptionHandling.InvalidParamException;
-import ulaval.glo2003.api.exceptionHandling.NotPermittedException;
+import ulaval.glo2003.domain.exceptions.InvalidParamException;
+import ulaval.glo2003.domain.exceptions.NotPermittedException;
 
 public class Product {
     private final String title;
