+++ src/main/java/ulaval/glo2003/domain/exceptions/MissingParamException.java
@@ -1,4 +1,7 @@
-package ulaval.glo2003.api.exceptionHandling;
+package ulaval.glo2003.domain.exceptions;
+
+import ulaval.glo2003.domain.exceptions.ErrorDescription;
+import ulaval.glo2003.domain.exceptions.ParamValidationException;
 
 public class MissingParamException extends ParamValidationException {
