+++ src/main/java/ulaval/glo2003/domain/exceptions/ParamValidationException.java
@@ -1,6 +1,7 @@
-package ulaval.glo2003.api.exceptionHandling;
+package ulaval.glo2003.domain.exceptions;
 
 import jakarta.validation.ValidationException;
+import ulaval.glo2003.domain.exceptions.ErrorDescription;
 
 public abstract class ParamValidationException extends ValidationException {
     public final ErrorDescription errorDescription;
